% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bar.R
\name{data_bar}
\alias{data_bar}
\title{Make a barplot for tabulated data}
\usage{
data_bar(x, i = 1, xlab = NULL, ylab = NULL, title = NULL, labels = NULL, ...)
}
\arguments{
\item{dwtable}{A dwtable object.}
}
\description{
Make a barplot for tabulated data
}
