% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bar.R
\name{data_groupmeans_bar}
\alias{data_groupmeans_bar}
\title{Make a barplot for group means}
\usage{
data_groupmeans_bar(x, xlab = NULL, ylab = NULL, title = NULL)
}
\arguments{
\item{x}{A dw_groupmeans data frame.}

\item{xlab}{A string to use for labelling the x axis}

\item{ylab}{A string to use for labelling the y axis}

\item{title}{A string to be used as the graph title}
}
\description{
Make a barplot for group means
}
