% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ci.R
\name{data_groupmeans_ci}
\alias{data_groupmeans_ci}
\title{Plot confidence intervals for group means}
\usage{
data_groupmeans_ci(x, xlab = NULL, ylab = NULL, title = NULL)
}
\arguments{
\item{x}{A dw_groupmeans data frame.}

\item{xlab}{A string to use for labelling the x axis}

\item{ylab}{A string to use for labelling the y axis}

\item{title}{A string to be used as the graph title}
}
\description{
Plot confidence intervals for group means
}
