% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pie.R
\name{data_pie}
\alias{data_pie}
\title{Make a pie chart for tabulated data}
\usage{
data_pie(x, i = 1, labels = NULL, title = NULL, values = NULL, ...)
}
\arguments{
\item{dwtable}{A dwtable object.}
}
\description{
Make a pie chart for tabulated data
}
