% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_histogram.R
\name{distribution_histogram}
\alias{distribution_histogram}
\alias{data_histogram}
\title{Make a histogram for untablulated data}
\usage{
distribution_histogram(x, xlab = NULL, title = NULL, values = NULL)
}
\arguments{
\item{x}{A vector.}

\item{xlab}{A string to use for labelling the x axis}

\item{title}{A string to be used as the graph title}

\item{values}{A vector of labels to use for labelling the slices.}
}
\description{
Make a histogram for untablulated data
}
