% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pie.R
\name{distribution_pie}
\alias{distribution_pie}
\alias{data_pie}
\title{Make a pie chart for tabulated data}
\usage{
distribution_pie(x, i = 1, title = NULL, values = NULL, ...)
}
\arguments{
\item{x}{A vector}

\item{i}{Numeric indicator of Which table from a list of
tables should be plotted.
Defaults to 1.}

\item{title}{A string to be used as the graph title}

\item{values}{A vector of labels to use for labelling the slices.
Must match the numerical order of the variable.}

\item{...}{Additional options}
}
\description{
Make a pie chart for tabulated data
}
