% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accesstemplates.R
\name{posterpdf}
\alias{posterpdf}
\title{create a copy of the ICORP pdf poster file}
\usage{
posterpdf(file)
}
\arguments{
\item{poster}{The name of the folder to be used which
is also the name of the file.}
}
\description{
Creates a folder and file with the skeleton file
}
