% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_image.R
\name{resize_image}
\alias{resize_image}
\title{Resize image for poster}
\usage{
resize_image(
  templatefolder = "myfolder",
  image = "myprofile_image",
  size = 0.5,
  type = c("profile", "qrcode", "other")
)
}
\arguments{
\item{templatefolder}{The name of the folder the template is in, which
contains the images folder holding images.}

\item{image}{The full file name of the image}

\item{size}{The size value to be used by \code{imager::imsize()}}

\item{type}{Choice of "profile", "qrcode" or "other". This determines
the name of the created resized image.}
}
\description{
Resize image for poster
}
