% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss1.R
\docType{data}
\name{rss1}
\alias{rss1}
\title{rss1 data set}
\format{
The RSS1 Data
}
\source{
\url{https://www.cdc.gov/nchs/rss/access.html#round1}
}
\usage{
rss1
}
\description{
Please review the information on the
source website and the rss1_codebook
article.
}
\keyword{datasets}
