% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss1_raw.R
\docType{data}
\name{rss1_raw}
\alias{rss1_raw}
\title{rss1_raw}
\format{
The RSS1 Data
}
\source{
\url{https://www.cdc.gov/nchs/rss/access.html#round1}
}
\usage{
rss1_raw
}
\description{
rss1_raw
}
\keyword{datasets}
