% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss2.R
\docType{data}
\name{rss2}
\alias{rss2}
\title{RSS2 data set}
\format{
The RSS2 Data
}
\source{
\url{https://www.cdc.gov/nchs/rss/access.html#round2}
}
\usage{
rss2
}
\description{
Please review the information on the
source website and the rss2_codebook
article.
}
\keyword{datasets}
