% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss2_raw.R
\docType{data}
\name{rss2_raw}
\alias{rss2_raw}
\title{rss2_raw}
\format{
The RSS2 Data
}
\source{
\url{https://www.cdc.gov/nchs/rss/access.html#round2}
}
\usage{
rss2_raw
}
\description{
rss2_raw
}
\keyword{datasets}
