% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss3.R
\docType{data}
\name{rss3}
\alias{rss3}
\title{rss3 data set}
\format{
The RSS3 Data
}
\source{
\url{https://www.cdc.gov/nchs/rss/access.html#round3}
}
\usage{
rss3
}
\description{
Please review the information on the
source website and the rss3_codebook
article.
}
\keyword{datasets}
