% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss3_raw.R
\docType{data}
\name{rss3_raw}
\alias{rss3_raw}
\title{rss3_raw}
\format{
The RSS3 Data
}
\source{
\url{https://www.cdc.gov/nchs/rss/access.html#round3}
}
\usage{
rss3_raw
}
\description{
rss3_raw
}
\keyword{datasets}
