% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss4.R
\docType{data}
\name{rss4}
\alias{rss4}
\title{rss4 data set
Please review the information on the
source website and the rss4_codebook
article.}
\format{
The RSS4 Data
}
\source{
\url{https://www.cdc.gov/nchs/rss/access.html#round4}
}
\usage{
rss4
}
\description{
rss4 data set
Please review the information on the
source website and the rss4_codebook
article.
}
\keyword{datasets}
