% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss4_raw.R
\docType{data}
\name{rss4_raw}
\alias{rss4_raw}
\title{rss4_raw}
\format{
The RSS4 Data
}
\source{
\url{https://www.cdc.gov/nchs/rss/access.html#round4}
}
\usage{
rss4_raw
}
\description{
rss4_raw
}
\keyword{datasets}
