% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss5.R
\docType{data}
\name{rss5}
\alias{rss5}
\title{rss5 data set}
\format{
The RSS5 Data
}
\source{
\url{https://www.cdc.gov/nchs/rss/access.html#round5}
}
\usage{
rss5
}
\description{
Please review the information on the
source website and the rss5_codebook
article.
}
\keyword{datasets}
