% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss5_raw.R
\docType{data}
\name{rss5_raw}
\alias{rss5_raw}
\title{rss5_raw}
\format{
The RSS5 Data
}
\source{
\url{https://www.cdc.gov/nchs/rss/access.html#round5}
}
\usage{
rss5_raw
}
\description{
rss5_raw
}
\keyword{datasets}
