% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss6.R
\docType{data}
\name{rss6}
\alias{rss6}
\title{rss6 data set}
\format{
The RSS6 Data
}
\source{
\url{https://www.cdc.gov/nchs/rapid-surveys/documentation/round-6.html}
}
\usage{
rss6
}
\description{
Please review the information on the
source website and the rss6_codebook
article.
}
\keyword{datasets}
