% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss7_raw.R
\docType{data}
\name{rss7_raw}
\alias{rss7_raw}
\title{rss7_raw}
\format{
The RSS7 Data
}
\source{
\url{https://www.cdc.gov/nchs/rapid-surveys/documentation/round-7.html}
}
\usage{
rss7_raw
}
\description{
rss7_raw
}
\keyword{datasets}
