% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accesstemplates.R
\name{basicplots}
\alias{basicplots}
\title{Create basic one variable plots}
\usage{
basicplots(project = "icorp26")
}
\arguments{
\item{project}{The name of the folder to be used which
is also the name of the project.
By default, srop2026..}
}
\description{
Creates a file that matches the basic plots vignette
}
