% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstarted.R
\name{create_starter}
\alias{create_starter}
\title{Creates a starter rmarkdown file}
\usage{
create_starter(filename)
}
\arguments{
\item{filename}{Name of the file to create}
}
\description{
Creates a starter rmarkdown file
}
