% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chisq.R
\name{data_chisq}
\alias{data_chisq}
\title{Calculate chi square test for tabulate}
\usage{
data_chisq(dwtable, ...)
}
\arguments{
\item{dwtable}{A dwtable object.}
}
\description{
Calculate chi square test for tabulate
}
