% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_histogram.R
\name{data_histogram}
\alias{data_histogram}
\title{Make a histogram for tablulated data}
\usage{
data_histogram(x, i = 1, xlab = NULL, title = NULL, values = NULL, ...)
}
\arguments{
\item{dwtable}{A dwtable object.}
}
\description{
Make a histogram for tablulated data
}
