% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bar.R
\name{distribution_bar}
\alias{distribution_bar}
\alias{data_bar}
\title{Make a barplot}
\usage{
distribution_bar(x, i = 1, xlab = NULL, ylab = NULL, title = NULL, ...)
}
\arguments{
\item{x}{A vector.}

\item{i}{Numeric indicator of Which table should be plotted.}

\item{xlab}{A string to use for labelling the x axis}

\item{ylab}{A string to use for labelling the y axis
Defaults to 1.}

\item{title}{A string to be used as the graph title}

\item{...}{Additional options}

\item{values}{A vector of labels to use for labelling the slices.
Must match the numerical order of the variable.}
}
\description{
Make a barplot
}
