% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicstatistics.R
\name{distribution_mean}
\alias{distribution_mean}
\title{calculate a mean}
\usage{
distribution_mean(x, remove_na = TRUE, convert_to_numeric = FALSE)
}
\arguments{
\item{x}{A vector}

\item{remove_na}{Logical indicating if NAs should be removed.
Defaults to TRUE}

\item{convert_to_numeric}{Logical indicating if non numeric vectors
should be coerced to numeric.}
}
\description{
calculate a mean
}
