% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicstatistics.R
\name{distribution_median}
\alias{distribution_median}
\title{Calculate a median}
\usage{
distribution_median(x, remove_na = TRUE, to_numeric = FALSE)
}
\arguments{
\item{x}{A vector}

\item{to_numeric}{Logical indicating if non numeric vectors
should be coerced to numeric.}
}
\description{
Calculate a median
}
