% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstarted.R
\name{getstarted}
\alias{getstarted}
\title{getstarted}
\usage{
getstarted(project = "srop2026")
}
\arguments{
\item{project}{The name of the folder to be used which
is also the name of the project.
By default, srop2026. You should only
run this once for a project.}
}
\description{
Creates a project with a few files to get started
}
