% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dw_crosstabs.R
\name{plot.datawizard_crosstab}
\alias{plot.datawizard_crosstab}
\title{Plot method for crosstabs}
\usage{
\method{plot}{datawizard_crosstab}(x, y, ...)
}
\arguments{
\item{x}{A datawizard_crosstabs object`.}

\item{y}{Not currently used}

\item{...}{Arguments to pass on. Currently accepts "title"
as the graph title.}
}
\value{
A ggplot2 representation of a datawizard
cross tabulation.
}
\description{
The \code{plot()} method for \code{dw_crosstabs} objects.
These objects are lists of \code{dw_crosstab}objects.
}
\details{
There is no standard way to plot a cross tabulation.
For these plot functions the format respond to the value
selected for the proportions argument of \code{data_tabulate()}
from the \code{datawizard} package. Those options are "row" "column" and "full"
along with NULL for no percentages (numbers only).

For "row" and "column" percentages, stacked bar plots of the respective
row and colum npercentages are displayed. For row percentages the bars
display horizontally

For tables with no percentage--raw numbers only--a heat map
describes the number of observations in each cell.

For tables with "all" percents, i.e. with values representing
the percentage of the whole sample the graph is a variation of a mosaic
plot where the size of each rectangle represents the proportion of the
whole sample falls in each cell.

To remove missing values from the plots, the data_tabulate
object should be created with the \code{remove_na} argument of TRUE.

The plots can be modified using ggplot2 syntax and options.
}
