% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dw_crosstabs.R
\name{plot.datawizard_crosstab}
\alias{plot.datawizard_crosstab}
\title{Plot method for crosstabs}
\usage{
\method{plot}{datawizard_crosstab}(x)
}
\arguments{
\item{x}{An object returned from \code{datawizard::datawizard_crosstab()}.}

\item{y}{Not currently used}

\item{...}{Not currently used}
}
\description{
The \code{plot()} method for \code{datawizard_crosstab} objects
}
