% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dw_crosstabs.R
\name{plot.datawizard_crosstabs}
\alias{plot.datawizard_crosstabs}
\title{Plot method for crosstabs}
\usage{
\method{plot}{datawizard_crosstabs}(x, y, ...)
}
\arguments{
\item{x}{An object returned from \code{datawizard::datawizard_crosstab()}.}

\item{y}{Not currently used}

\item{...}{Additional options. a \code{title} attribute will
make an overall title for the graph}
}
\description{
The \code{plot()} method for \code{datawizard_crosstab} objects
which are lists.
}
