% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dw_groupmeans.R
\name{plot.dw_groupmeans}
\alias{plot.dw_groupmeans}
\title{Plot confidence intervals for group means}
\usage{
\method{plot}{dw_groupmeans}(x, title = NULL, caption = TRUE)
}
\arguments{
\item{x}{A dw_groupmeans data frame.}

\item{title}{A string to be used as the graph title}

\item{caption}{Logical indicating if a caption summarizing
the anova results should be included.}
}
\description{
Plot confidence intervals for group means
}
