% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss7.R
\docType{data}
\name{rss7}
\alias{rss7}
\title{rss7}
\format{
The RSS7 Data
}
\source{
\url{https://www.cdc.gov/nchs/rapid-surveys/documentation/round-7.html}
}
\usage{
rss7
}
\description{
rss7
}
\keyword{datasets}
